UBUNTU: SAUCE: clocksource: hyper-v: do not use an insanely big TSC in hv_read_tsc_page_tsc()
When vCPU0 starts on an AMD CPU, the hypervisor initializes the TSC-offset
field in vCPU0's VMCB using the opposite number of the TSC of the physical
CPU, on which vCPU0 is running. When vCPU0 executes RDTSC, the returned
TSC reading is the TSC-offset + the physical TSC.

When the other vCPUs starts, the hypervisor uses the same TSC-offset value to
initialize the TSC-offset fields in their VMCBs.

It's been reported that: in the early boot code of the non-boot vCPUs, very
rarely the raw TSC reading from RDTSC can be so big that it's actually a
negative s64 integer.

This happens probably because the TSCs of the physical CPUs might drift a
little bit, i.e. if vCPU1's physical CPU's TSC is a little smaller than
vCPU0's physical CPU's TSC, RDTSC in vCPU1's early boot code can return a
"negative" TSC reading, causing an insanely big time value to be returned
from the Hyper-V TSC page; if the insanely big time value is used to
program the Hyper-V timer, the timer can only fire in hundreds of years,
confusing the scheduler on the vCPU and resulting in stuck processes on
the vCPU.

A "negative" TSC reading can also cause some kernel messages prefixed by
insanely big timestamps.

While we're trying to understand how exactly the "negative" TSC is returned
from RDTSC, let's work around the issue by avoiding using such an insane
TSC reading in hv_read_tsc_page_tsc().

Note: when vCPU1's physical CPU's TSC grows big enough, RDTSC on vCPU1
will no longer return "negative" values.

Note: it looks like the issue doesn't reproduce if the Invariant TSC is
supported and used, so there is no need to check the TSC reading in
the TSC clocksource's code.

Signed-off-by: Dexuan Cui <decui@microsoft.com>